fix: update mat4 submat array indexing
this change causes the tests to pass even though I thought this
causes incorrect indexing into the array?

LLMs agree that this is the correct way to index into a column major
matrix.

This change appears to cause the submatrix to also be in column
major format, but the tests still pass...