feat: add cubic bezier and hermite splines
spline curves are used to describe the change in a value between keyframes in an animation track. The two common types
of spline curves are cubic bezier and hermite. As such, an
implementation of the two curves is provided to help the user
understand them. An example would be spline interpolation between
two vector 3's for the position of a bone.