feat: create animation data structures and functions (#13 via #14)
* feat: add cubic bezier and hermite splines

spline curves are used to describe the change in a value between keyframes in an animation track. The two common types
of spline curves are cubic bezier and hermite. As such, an
implementation of the two curves is provided to help the user
understand them. An example would be spline interpolation between
two vector 3's for the position of a bone.

* build: add curve file to project

* build: add interp type enum file to project

* feat: create keyframe interp type enum

* docs: explain anim namespace purpose

* feat: create animation frame type

* feat: define animation track type

* build: add animation files to project

* feat: add helper functions for track impl

* docs: improve function param name

new name better describes that the float array pointer is taken from
the frame's data array

* feat: implement track functions

* misc: create SRT track file

* build: add SRT track header file to project

* feat: create SRTtrack header

* docs: document several SRTtrack functions

* feat: implement SRTtrack getters

* build: add SRTtrack impl file to project

* feat: impl SRTtrack start and end time getters

* feat: implement SRTtrack sampling function