fix: resize anim track to correct size
before the anim track, which stores the frames for one aspect of a
bone (i.e. rotation track: stores rotation frames, for one bone), was
being resized to the number of components in a frame which is
incorrect. It is now resized to the number of frames in the track